requirements.txt
anthropic>=0.71.0
python-dotenv>=1.0.0
